+++ app/src/main/AndroidManifest.xml
@@ -11,7 +11,6 @@
 		android:supportsRtl="true"
 		android:theme="@style/Theme.AppCongreso"
 		tools:ignore="AllowBackup">
-		<activity android:name=".ui.login.HomeActivity"></activity>
 		<activity
 			android:name=".MainActivity"
 			android:label="@string/app_name"
