+++ app/src/main/res/navigation/mobile_navigation.xml
@@ -20,13 +20,23 @@
 		android:id="@+id/nav_lecture_list"
 		android:name="org.grupotres.appcongreso.ui.lectures.LectureListFragment"
 		android:label="@string/menu_lecture_list"
-		tools:layout="@layout/fragment_lecture_list" />
+		tools:layout="@layout/fragment_lecture_list" >
+
+		<action
+			android:id="@+id/action_home"
+			app:destination="@id/nav_lecture_list"/>
+	</fragment>
 
 	<fragment
 		android:id="@+id/nav_speaker_list"
 		android:name="org.grupotres.appcongreso.ui.speakers.SpeakerListFragment"
 		android:label="@string/menu_speaker_list"
-		tools:layout="@layout/fragment_speaker_list" />
+		tools:layout="@layout/fragment_speaker_list">
+
+		<action
+			android:id="@+id/action_home"
+			app:destination="@id/nav_speaker_list"/>
+	</fragment>
 
 	<fragment
 		android:id="@+id/nav_resources"
